Move `relay/http/streams` to `relay/client/streams`
